# feat(python): guide cython to optimize code generation